REFACTOR
- seperated everything in cli and core module
- made poms pretty
- made subfolder for kotlin project
- readme went in main folder
