+++ PasswordManager/password-core/src/main/resources/init-sqlite-data.sql
